Update install script

+++ curl_vimtools.sh
@@ -1,8 +1,18 @@
 LIGHT_PURPLE='\033[1;35m'
-NO_COLOR='\033[0m' # No Color
+LIGHT_RED='\033[1;31m'
+LIGHT_YELLOW='\033[1;33m'
+NO_COLOR='\033[0m'
 
 echo_step() {
     echo -e "${LIGHT_PURPLE}$1${NO_COLOR}"
+}
+
+echo_error() {
+    echo -e "${LIGHT_RED}ERROR: $1${NO_COLOR}"
+}
+
+echo_warning() {
+    echo -e "${LIGHT_YELLOW}WARNING: $1${NO_COLOR}"
 }
 
 fetch_vimtools() {
@@ -12,30 +22,33 @@ fetch_vimtools() {
     local inflated="${repo}-${branch}"
     local vim_directory=".vim"
 
-    echo_step "Step 1. Downloading ${archive}..."
-    curl -L https://github.com/dsw7/${repo}/archive/${branch}.zip --output $archive
-    echo ""
+    echo_step "[Step 1] - Downloading ${archive}..."
+    curl -L https://github.com/dsw7/${repo}/archive/${branch}.zip --output $archive --fail
+    if [ $? -ne 0 ]
+    then
+        echo_error "Failed to fetch VimTools!"
+        return
+    fi
+    echo
 
-    echo_step "Step 2. Inflating ${archive}..."
+    echo_step "[Step 2] - Inflating ${archive}..."
     unzip -o ${archive}
-    echo ""
+    echo
 
-    echo_step "Step 3. Remove existing $vim_directory directory..."
+    echo_step "[Step 3] - Remove existing $vim_directory directory..."
     if [ -d $vim_directory ]
     then
         rm -rfv $vim_directory
     else
-        echo "No existing $vim_directory directory found!"
+        echo_warning "No existing $vim_directory directory found!"
     fi
-    echo ""
+    echo
 
-    echo_step "Step 4. Rename inflated directory..."
+    echo_step "[Step 4] - Rename inflated directory..."
     mv -v $inflated $vim_directory
-    echo ""
+    echo
 
-    echo_step "Step 5. Clean up any remaining files..."
+    echo_step "[Step 5] - Clean up any remaining files..."
     rm -v $archive
-    echo ""
-
-    echo "Complete!"
+    echo
 }
